Teams detail page was starting to get crowded. This is the start of a mostly visual, but also structural, refactor.
