+++ src/teams/templates/info_item_form.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'team_base.html' %}
 {% load commonmark %}
 {% load bootstrap3 %}
 
@@ -11,7 +11,7 @@ Create Info item
 in "{{ form.instance.category.headline }}"
 {% endblock %}
 
-{% block content %}
+{% block team_content %}
 <div class="panel panel-default">
     <div class="panel-heading">
         <h4>
@@ -36,6 +36,6 @@ in "{{ form.instance.category.headline }}"
             {% endif %}
         </form>
     </div>
-    <div class="panel-footer"><i>This info item belongs to the <a href="{% url 'teams:detail' team_slug=team.slug camp_slug=team.camp.slug %}">{{ team.name }} Team</a></i></div>
+    <div class="panel-footer"><i>This info item belongs to the <a href="{% url 'teams:general' team_slug=team.slug camp_slug=team.camp.slug %}">{{ team.name }} Team</a></i></div>
 </div>
 {% endblock %}
