+++ src/teams/templates/task_form.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'team_base.html' %}
 {% load commonmark %}
 {% load bootstrap3 %}
 
@@ -11,7 +11,7 @@ Create Task
 for {{ team.name }} Team
 {% endblock %}
 
-{% block content %}
+{% block team_content %}
 <div class="panel panel-default">
     <div class="panel-heading">
         <h4>
@@ -30,7 +30,7 @@ for {{ team.name }} Team
             <button type="submit" class="btn btn-primary">{% if form.instance.id %}Save{% else %}Create{% endif %}</button>
         </form>
     </div>
-    <div class="panel-footer"><i>This task belongs to the <a href="{% url 'teams:detail' team_slug=team.slug camp_slug=team.camp.slug %}">{{ team.name }} Team</a></i></div>
+    <div class="panel-footer"><i>This task belongs to the <a href="{% url 'teams:general' team_slug=team.slug camp_slug=team.camp.slug %}">{{ team.name }} Team</a></i></div>
 </div>
 {% endblock %}
