+++ src/teams/templates/team_info_categories.html
@@ -1,0 +1,52 @@
+{% extends 'team_base.html' %}
+{% load commonmark %}
+{% load bootstrap3 %}
+{% load teams_tags %}
+
+
+{% block team_content %}
+
+{% if request.user in team.responsible_members.all and team.info_categories.exists %}
+<h4> SHOULD NOT HAPPEN !!! </h4>
+{% endif %}
+
+<div class="panel panel-default">
+    <div class="panel-heading">
+        <h4>Info Categories</h4>
+    </div>
+    <div class="panel-body">
+
+        {% for info_category in team.info_categories.all %}
+
+            <h4>{{ info_category.headline }}</h4>
+            <table class="table table-hover">
+                <thead>
+                    <tr>
+                        <th>Item name</th>
+                        <th>Action</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {% for item in info_category.infoitems.all %}
+                    <tr>
+                        <td>{{ item.headline }}</td>
+                        <td>
+                            <a href="{% url 'teams:info_item_update' camp_slug=camp.slug team_slug=team.slug category_anchor=info_category.anchor item_anchor=item.anchor %}"
+                               class="btn btn-primary btn-sm">
+                                <i class="fas fa-edit"></i> Edit
+                            </a>
+                        </td>
+                    </tr>
+                    {% endfor %}
+                </tbody>
+            </table>
+
+            <a href="{% url 'teams:info_item_create' camp_slug=camp.slug team_slug=team.slug category_anchor=info_category.anchor %}" class="btn btn-primary"><i class="fas fa-plus"></i> Create Info Item</a>
+
+            <hr />
+
+        {% endfor %}
+    </div>
+</div>
+
+{% endblock %}
