+++ src/teams/templates/team_leave.html
@@ -1,11 +1,11 @@
-{% extends 'base.html' %}
+{% extends 'team_base.html' %}
 {% load commonmark %}
 
 {% block title %}
 Leave Team: {{ team.name }} | {{ block.super }}
 {% endblock %}
 
-{% block content %}
+{% block team_content %}
 
 <h3>Leave {{ team.name }} Team</h3>
 <p class="lead">Really leave the <b>{{ team.name }}</b> team?<p>
