+++ src/teams/templates/team_tasks.html
@@ -1,0 +1,42 @@
+{% extends 'team_base.html' %}
+{% load commonmark %}
+{% load bootstrap3 %}
+{% load teams_tags %}
+
+{% block team_content %}
+
+<div class="panel panel-default">
+    <div class="panel-heading">
+        <h4>Tasks</h4>
+    </div>
+    <div class="panel-body">
+        <p>The {{ team.name }} Team is responsible for the following tasks</p>
+        <table class="table table-hover">
+            <thead>
+            <tr>
+                <th>Name</th>
+                <th>Description</th>
+                <th>Action</th>
+            </tr>
+            </thead>
+            <tbody>
+            {% for task in team.tasks.all %}
+            <tr>
+                <td><a href="{% url 'teams:task_detail' slug=task.slug camp_slug=camp.slug team_slug=team.slug %}">{{ task.name }}</a></td>
+                <td>{{ task.description }}</td>
+                <td>
+                    <a href="{% url 'teams:task_detail' camp_slug=camp.slug team_slug=team.slug slug=task.slug %}" class="btn btn-primary btn-sm"><i class="fas fa-search"></i> Details</a>
+                    {% if request.user in team.responsible_members.all %}
+                    <a href="{% url 'teams:task_update' camp_slug=camp.slug team_slug=team.slug slug=task.slug %}" class="btn btn-primary btn-sm"><i class="fas fa-edit"></i> Edit Task</a>
+                    {% endif %}
+                </td>
+            </tr>
+            {% endfor %}
+            </tbody>
+        </table>
+        {% if request.user in team.responsible_members.all %}
+        <a href="{% url 'teams:task_create' camp_slug=camp.slug team_slug=team.slug %}" class="btn btn-primary"><i class="fas fa-plus"></i> Create Task</a>
+        {% endif %}
+    </div>
+</div>
+{% endblock %}
