+++ src/teams/templates/teammember_approve.html
@@ -1,11 +1,11 @@
-{% extends 'base.html' %}
+{% extends 'team_base.html' %}
 {% load commonmark %}
 
 {% block title %}
 Approve team member {{ teammember.user.profile.name }} for the {{ teammember.team.name }} team
 {% endblock %}
 
-{% block content %}
+{% block team_content %}
 
 <h3>Approve member {{ teammember.user.profile.name }} for the {{ teammember.team.name }} team</h3>
 <p class="lead">Really approve the user <b>{{ teammember.user.profile.name }}</b> for the {{ teammember.team.name }} team? The user will receive an email with a message.<p>
@@ -13,6 +13,7 @@ Approve team member {{ teammember.user.profile.name }} for the {{ teammember.tea
 {% csrf_token %}
 {{ form }}
 <button class="btn btn-success" type="submit"><i class="fas fa-check"></i> Add teammember</button>
-<a href="{% url 'teams:detail' camp_slug=teammember.team.camp.slug team_slug=teammember.team.slug %}" class="btn btn-default" type="submit"><i class="fas fa-times"></i> Cancel</a>
+<a href="{% url 'teams:general' camp_slug=teammember.team.camp.slug team_slug=teammember.team.slug %}" class="btn btn-default" type="submit"><i class="fas fa-times"></i> Cancel</a>
 </form>
+
 {% endblock %}
