+++ src/templates/base.html
@@ -58,7 +58,7 @@
             </a>
             {% endif %}
         </div>
-        <div id="navbar" class="navbar-collapse collapse">
+        <div id="top-navbar" class="navbar-collapse collapse">
           <ul class="nav navbar-nav">
             <li><a href="{% url 'news:index' %}">News</a></li>
             <li><a href="{% url 'shop:index' %}">Shop</a></li>
